+++ .ci-pipelines/ci-testing-script.sh
@@ -12,8 +12,12 @@ cwd=$(pwd -P)
 
 # Run C-I tests with various mechanism + integrator combinations
 for this_test in ${GENERAL_TESTS}; do
-    run_ci_test "${this_test}" "${cwd}"
+    run_ci_test "${this_test}" "${cwd}" ""
 done
+
+# Run the MCM test separately
+extra="EXTERNAL_RATES_F90=constants_mcm.f90 EXTERNAL_RATES_OBJ=constants_mcm.o"
+run_ci_test "${MCM_TEST}" "${cwd}" "${extra}"
 
 # Run a C-I test to see if the #MINVERSION command works as advertised
 run_minversion_ci_test "${MINVERSION_TEST}" "${cwd}"