+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - TBD
 ### Removed
 - `TRANSPORT` and `TRANSPORTALL` input options; these were obsolete
+- `LUMP` input option; this was obsolete
 
 ## [3.0.2] - 2023-06-02
 ### Added