+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - TBD
 ### Added
 - `#AUTOREDUCE` has been added to the list of KPP commands in the ReadTheDocs documentaton
+- Added `examples/mcm` folder with minimal example for the Master Chemical Mechanism
+- Added C-I test for MCM, based on the minimal example
 
 ### Removed
 - `TRANSPORT` and `TRANSPORTALL` input options; these were obsolete